Add connectTerms property to TriggerGateController
